refactor(validation): delegate Cars validation to utils.Validator
